# api-openapi.yaml
openapi: "3.0.3"
info:
    title: "Cocomine API"
    description: "HTTP & WS endpoints for Cocomine VPN manager"
    version: "1.0"
servers:
    -   url: "https://api.cocomine.cc"
        description: "Production"
    -   url: "http://localhost:8088"
        description: "Local"
components:
    securitySchemes:
        CfAccessHeader:
            type: apiKey
            in: header
            name: Cf-Access-Jwt-Assertion
            description: "Cloudflare Access JWT (web)."
        CfAccessCookie:
            type: apiKey
            in: cookie
            name: CF_Authorization
            description: "Cloudflare Access JWT (web cookie)."
        AppCheck:
            type: http
            scheme: bearer
            bearerFormat: "Firebase App Check"
        CfBearer:
            type: http
            scheme: bearer
            bearerFormat: "Cloudflare Access"
    schemas:
        Error:
            type: object
            description: "Standard error envelope"
            properties:
                code:
                    type: integer
                    description: "Application status code"
                message:
                    type: string
                    description: "Human-readable error message"
...
        UserInfo:
            type: object
            description: "Claims resolved from CF Access"
            properties:
                email:
                    type: string
                    description: "User email"
                    example: "example@example.com"
                name:
                    type: string
                    nullable: true
                    description: "Display name"
                    example: "John Doe"
                aud:
                    oneOf:
                        -   type: array
                            items: { type: string, example: "xxxx" }
                        -   type: string
                            example: "xxxx"
                    description: "Audience claim"
                sub:
                    type: string
                    nullable: true
                    description: "Subject claim"
                    example: "xxxx"
        TokenResponse:
            type: object
            description: "App JWT response"
            properties:
                code:
                    type: integer
                    description: "Application status code"
                    example: 200
                data:
                    type: object
                    description: "Payload wrapper"
                    properties:
                        token:
                            type: string
                            description: "Signed JWT"
                            example: "Signed jwt token"
...
    examples:
        ErrorExpired:
            summary: Token expired
            description: "Expired token"
            value:
                code: 401
                message: "Token Expired"
        ErrorMissing:
            summary: Missing token
            value:
                code: 401
                message: "Missing required cf authorization token"
        ErrorInvalid:
            summary: Invalid token
            description: "Invalid token"
            value:
                code: 403
                message: "Invalid Token"
...
paths:
...
    /auth/odic/exchange:
        get:
            summary: Exchange CF token for app JWT
            description: "Validate CF Access token then issue app JWT"
            security:
                -   CfBearer: [ ]
            responses:
                '200':
                    description: JWT issued
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TokenResponse'
                '401':
                    description: Missing/expired token
                    content:
                        application/json:
                            examples:
                                expired:
                                    $ref: '#/components/examples/ErrorExpired'
                                missing:
                                    $ref: '#/components/examples/ErrorMissing'
                            schema:
                                $ref: '#/components/schemas/Error'
                '403':
                    description: Token invalid
                    content:
                        application/json:
                            examples:
                                invalid:
                                    $ref: '#/components/examples/ErrorInvalid'
                            schema:
                                $ref: '#/components/schemas/Error'
                '500':
                    description: Internal error
                    content:
                        application/json:
                            example:
                                code: 500
                                message: "Internal server error"
                            schema:
                                $ref: '#/components/schemas/Error'
    #OK
    /userinfo:
        get:
            summary: Get authenticated user info
            description: "Returns decoded CF Access claims + profileType name when available"
            security:
                -   CfAccessHeader: [ ]
                -   CfAccessCookie: [ ]
            responses:
                '200':
                    description: User payload
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserInfo'
                '401':
                    description: Missing/expired token
                    content:
                        application/json:
                            examples:
                                expired:
                                    $ref: '#/components/examples/ErrorExpired'
                                missing:
                                    $ref: '#/components/examples/ErrorMissing'
                            schema:
                                $ref: '#/components/schemas/Error'
                '403':
                    description: Invalid token
                    content:
                        application/json:
                            examples:
                                invalid:
                                    $ref: '#/components/examples/ErrorInvalid'
                            schema:
                                $ref: '#/components/schemas/Error'
                '500':
                    description: Internal error
                    content:
                        application/json:
                            example:
                                code: 500
                                message: "Internal server error"
                            schema:
                                $ref: '#/components/schemas/Error'
    #OK
...